% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_url}
\alias{check_url}
\title{Check That A Webpage Exists In The Wayback Machine}
\usage{
check_url(url, message = TRUE)
}
\arguments{
\item{url}{A string containing the URL to the webpage.}

\item{message}{Print friendly messages to the console? Default is \code{TRUE}.}
}
\value{
If request is successful (i.e. the request was successfully performed
and a response with HTTP status code <400 was recieved), an HTTP response;
otherwise throws an error.
}
\description{
Check That A Webpage Exists In The Wayback Machine
}
\examples{
# Check that the CRAN homepage is in the Wayback Machine:
check_url("https://cran.r-project.org/")
}
