% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchive-package.R
\docType{package}
\name{rchive-package}
\alias{rchive}
\alias{rchive-package}
\title{rchive: A Simple Interface For The Internet Archive}
\description{
Check if a webpage exists in the Internet Archive's Wayback Machine, and add it if it doesn't.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/rchive}
  \item Report bugs at \url{https://github.com/lddurbin/rchive/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com}

}
\keyword{internal}
