% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_url}
\alias{check_url}
\title{Check That Webpage Exists In Wayback Machine}
\usage{
check_url(url)
}
\arguments{
\item{url}{A string containing the URL to the webpage.}
}
\value{
If request is successful (i.e. the request was successfully performed
and a response with HTTP status code <400 was recieved), an HTTP response;
otherwise throws an error.
}
\description{
Check That Webpage Exists In Wayback Machine
}
\examples{
# Check that the CRAN homepage is in the Wayback Machine:
check_url("https://cran.r-project.org/")
}
